pyproject.toml
[project]
name = "scalping-strategy"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2",
    "numpy>=2.3.4",
    "openai>=2.7.1",
    "pandas>=2.3.3",
    "pydantic-ai>=1.12.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
    "v20>=3.0.25.0",
]
